refactor: decouple config read from lib.rs
